Moving queries to common/queries (#4248)
