+++ components/shopify_partner/common/queries/getAppInstalls.mjs
