+++ components/shopify_partner/sources/new-app-relationship-events/new-app-relationship-events.mjs
@@ -1,5 +1,5 @@
 import common from "../../common/common.mjs";
-import shopify from "../../shopify_partner.app.mjs";
+import shopify from "../../common/shopify_partner.app.mjs";
 import getAppRelationshipEventsBackwards from "../../queries/getAppRelationshipEventsBackwards.mjs";
 import getAppRelationshipEventsForwards from "../../queries/getAppRelationshipEventsForwards.mjs";
 
@@ -7,7 +7,7 @@ export default {
   key: "shopify_partner-new-app-relationship-events",
   name: "New App Relationship Events",
   type: "source",
-  version: "0.0.6",
+  version: "0.0.7",
   description: "Emit new events when new shops installs, uninstalls, subscribes or unsubscribes your app.",
   ...common,
   props: {
